+++ cmd/influx/delete.go
@@ -12,7 +12,7 @@ import (
 var deleteFlags http.DeleteRequest
 
 func cmdDelete(f *globalFlags, opt genericCLIOpts) *cobra.Command {
-	cmd := opt.newCmd("delete", fluxDeleteF)
+	cmd := opt.newCmd("delete", fluxDeleteF, true)
 	cmd.Short = "Delete points from influxDB"
 	cmd.Long = `Delete points from influxDB, by specify start, end time
 	and a sql like predicate string.`